Get rid of mailer concern - it did not play well with the sort-of-singleton magic around mailer instance/class methods
