+++ plugins/ShinyCMS/.rubocop.yml
@@ -43,7 +43,6 @@ Metrics/BlockLength:
     - ../../spec/mailer_previews/*_spec.rb
     - ../Shiny*/shiny*.gemspec
     - ../Shiny*/spec/**/*.rb
-    - ../ShinyCMS/app/mailers/concerns/shinycms/mailer.rb
     - ../ShinyCMS/app/models/concerns/shinycms/*.rb
 
 # Spec files with I18n.t calls have long lines. I can live with that.
